+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/pyaud-plugins/compare/v0.20.0...HEAD)
+[Unreleased](https://github.com/jshwi/pyaud-plugins/compare/v0.21.0...HEAD)
+------------------------------------------------------------------------
+
+[0.21.0](https://github.com/jshwi/pyaud-plugins/releases/tag/v0.21.0) - 2024-01-31
 ------------------------------------------------------------------------
 ### Security
 - bump dependencies including extra style enforcement